reworked vespa indexing
