+++ backend/danswer/document_index/vespa/index.py
@@ -1,69 +1,59 @@
 import concurrent.futures
 import io
-import json
 import os
-import string
 import time
 import zipfile
-from collections.abc import Callable
-from collections.abc import Mapping
 from dataclasses import dataclass
 from datetime import datetime
 from datetime import timedelta
-from datetime import timezone
-from typing import Any
 from typing import BinaryIO
-from typing import cast
 
 import httpx
 import requests
-from retry import retry
 
-from danswer.configs.app_configs import LOG_VESPA_TIMING_INFORMATION
-from danswer.configs.app_configs import VESPA_CONFIG_SERVER_HOST
-from danswer.configs.app_configs import VESPA_HOST
-from danswer.configs.app_configs import VESPA_PORT
-from danswer.configs.app_configs import VESPA_TENANT_PORT
 from danswer.configs.chat_configs import DOC_TIME_DECAY
 from danswer.configs.chat_configs import EDIT_KEYWORD_QUERY
 from danswer.configs.chat_configs import HYBRID_ALPHA
 from danswer.configs.chat_configs import NUM_RETURNED_HITS
 from danswer.configs.chat_configs import TITLE_CONTENT_RATIO
-from danswer.configs.constants import ACCESS_CONTROL_LIST
-from danswer.configs.constants import BLURB
-from danswer.configs.constants import BOOST
-from danswer.configs.constants import CHUNK_ID
-from danswer.configs.constants import CONTENT
-from danswer.configs.constants import DOC_UPDATED_AT
-from danswer.configs.constants import DOCUMENT_ID
-from danswer.configs.constants import DOCUMENT_SETS
-from danswer.configs.constants import EMBEDDINGS
-from danswer.configs.constants import HIDDEN
-from danswer.configs.constants import INDEX_SEPARATOR
-from danswer.configs.constants import MEGA_CHUNK_REFERENCE_IDS
-from danswer.configs.constants import METADATA
-from danswer.configs.constants import METADATA_LIST
-from danswer.configs.constants import METADATA_SUFFIX
-from danswer.configs.constants import PRIMARY_OWNERS
-from danswer.configs.constants import RECENCY_BIAS
-from danswer.configs.constants import SECONDARY_OWNERS
-from danswer.configs.constants import SECTION_CONTINUATION
-from danswer.configs.constants import SEMANTIC_IDENTIFIER
-from danswer.configs.constants import SKIP_TITLE_EMBEDDING
-from danswer.configs.constants import SOURCE_LINKS
-from danswer.configs.constants import SOURCE_TYPE
-from danswer.configs.constants import TITLE
-from danswer.configs.constants import TITLE_EMBEDDING
 from danswer.configs.model_configs import SEARCH_DISTANCE_CUTOFF
-from danswer.connectors.cross_connector_utils.miscellaneous_utils import (
-    get_experts_stores_representations,
-)
-from danswer.document_index.document_index_utils import get_uuid_from_chunk
 from danswer.document_index.interfaces import DocumentIndex
 from danswer.document_index.interfaces import DocumentInsertionRecord
 from danswer.document_index.interfaces import UpdateRequest
-from danswer.document_index.vespa.utils import remove_invalid_unicode_chars
+from danswer.document_index.interfaces import VespaChunkRequest
+from danswer.document_index.vespa.chunk_retrieval_utils import (
+    get_all_vespa_ids_for_document_id,
+)
+from danswer.document_index.vespa.chunk_retrieval_utils import manage_batch_retrieval
+from danswer.document_index.vespa.chunk_retrieval_utils import (
+    parallel_visit_api_retrieval,
+)
+from danswer.document_index.vespa.chunk_retrieval_utils import query_vespa
+from danswer.document_index.vespa.deletion_utils import delete_vespa_docs
+from danswer.document_index.vespa.indexing_utils import batch_index_vespa_chunks
+from danswer.document_index.vespa.indexing_utils import clean_chunk_id_copy
+from danswer.document_index.vespa.indexing_utils import (
+    get_existing_documents_from_chunks,
+)
+from danswer.document_index.vespa.shared_utils.vespa_request_builders import (
+    build_vespa_filters,
+)
 from danswer.document_index.vespa.utils import replace_invalid_doc_id_characters
+from danswer.document_index.vespa_constants import ACCESS_CONTROL_LIST
+from danswer.document_index.vespa_constants import BATCH_SIZE
+from danswer.document_index.vespa_constants import BOOST
+from danswer.document_index.vespa_constants import CONTENT_SUMMARY
+from danswer.document_index.vespa_constants import DANSWER_CHUNK_REPLACEMENT_PAT
+from danswer.document_index.vespa_constants import DATE_REPLACEMENT
+from danswer.document_index.vespa_constants import DOCUMENT_ID_ENDPOINT
+from danswer.document_index.vespa_constants import DOCUMENT_REPLACEMENT_PAT
+from danswer.document_index.vespa_constants import DOCUMENT_SETS
+from danswer.document_index.vespa_constants import HIDDEN
+from danswer.document_index.vespa_constants import NUM_THREADS
+from danswer.document_index.vespa_constants import VESPA_APPLICATION_ENDPOINT
+from danswer.document_index.vespa_constants import VESPA_DIM_REPLACEMENT_PAT
+from danswer.document_index.vespa_constants import VESPA_TIMEOUT
+from danswer.document_index.vespa_constants import YQL_BASE
 from danswer.indexing.models import DocMetadataAwareIndexChunk
 from danswer.search.models import IndexFilters
 from danswer.search.models import InferenceChunkUncleaned
@@ -71,39 +61,9 @@ from danswer.search.retrieval.search_runner import query_processing
 from danswer.search.retrieval.search_runner import remove_stop_words_and_punctuation
 from danswer.utils.batching import batch_generator
 from danswer.utils.logger import setup_logger
-from danswer.utils.threadpool_concurrency import run_functions_tuples_in_parallel
 from shared_configs.model_server_models import Embedding
 
 logger = setup_logger()
-
-VESPA_DIM_REPLACEMENT_PAT = "VARIABLE_DIM"
-DANSWER_CHUNK_REPLACEMENT_PAT = "DANSWER_CHUNK_NAME"
-DOCUMENT_REPLACEMENT_PAT = "DOCUMENT_REPLACEMENT"
-DATE_REPLACEMENT = "DATE_REPLACEMENT"
-
-# config server
-VESPA_CONFIG_SERVER_URL = f"http://{VESPA_CONFIG_SERVER_HOST}:{VESPA_TENANT_PORT}"
-VESPA_APPLICATION_ENDPOINT = f"{VESPA_CONFIG_SERVER_URL}/application/v2"
-
-# main search application
-VESPA_APP_CONTAINER_URL = f"http://{VESPA_HOST}:{VESPA_PORT}"
-# danswer_chunk below is defined in vespa/app_configs/schemas/danswer_chunk.sd
-DOCUMENT_ID_ENDPOINT = (
-    f"{VESPA_APP_CONTAINER_URL}/document/v1/default/{{index_name}}/docid"
-)
-SEARCH_ENDPOINT = f"{VESPA_APP_CONTAINER_URL}/search/"
-
-_MAX_ID_SEARCH_QUERY_SIZE = 400
-_BATCH_SIZE = 128  # Specific to Vespa
-_NUM_THREADS = (
-    32  # since Vespa doesn't allow batching of inserts / updates, we use threads
-)
-# up from 500ms for now, since we've seen quite a few timeouts
-# in the long term, we are looking to improve the performance of Vespa
-# so that we can bring this back to default
-_VESPA_TIMEOUT = "3s"
-# Specific to Vespa, needed for highlighting matching keywords / section
-CONTENT_SUMMARY = "content_summary"
 
 
 @dataclass
@@ -111,711 +71,6 @@ class _VespaUpdateRequest:
     document_id: str
     url: str
     update_request: dict[str, dict]
-
-
-class VespaChunkRequest:
-    document_id: str
-    min_chunk_ind: int | None
-    max_chunk_ind: int | None
-
-
-@retry(tries=3, delay=1, backoff=2)
-def _does_document_exist(
-    doc_chunk_id: str,
-    index_name: str,
-    http_client: httpx.Client,
-) -> bool:
-    """Returns whether the document already exists and the users/group whitelists
-    Specifically in this case, document refers to a vespa document which is equivalent to a Danswer
-    chunk. This checks for whether the chunk exists already in the index"""
-    doc_url = f"{DOCUMENT_ID_ENDPOINT.format(index_name=index_name)}/{doc_chunk_id}"
-    doc_fetch_response = http_client.get(doc_url)
-
-    if doc_fetch_response.status_code == 404:
-        return False
-
-    if doc_fetch_response.status_code != 200:
-        logger.debug(f"Failed to check for document with URL {doc_url}")
-        raise RuntimeError(
-            f"Unexpected fetch document by ID value from Vespa "
-            f"with error {doc_fetch_response.status_code}"
-        )
-    return True
-
-
-def _vespa_get_updated_at_attribute(t: datetime | None) -> int | None:
-    if not t:
-        return None
-
-    if t.tzinfo != timezone.utc:
-        raise ValueError("Connectors must provide document update time in UTC")
-
-    return int(t.timestamp())
-
-
-def _get_vespa_chunks_by_document_id(
-    chunk_request: VespaChunkRequest,
-    index_name: str,
-    get_mega_chunks: bool = False,
-    user_access_control_list: list[str] | None = None,
-    field_names: list[str] | None = None,
-) -> list[dict]:
-    # Constructing the URL for the Visit API
-    # NOTE: visit API uses the same URL as the document API, but with different params
-    url = DOCUMENT_ID_ENDPOINT.format(index_name=index_name)
-
-    # build the list of fields to retrieve
-    field_set_list = (
-        None
-        if not field_names
-        else [f"{index_name}:{field_name}" for field_name in field_names]
-    )
-    acl_fieldset_entry = f"{index_name}:{ACCESS_CONTROL_LIST}"
-    if (
-        field_set_list
-        and user_access_control_list
-        and acl_fieldset_entry not in field_set_list
-    ):
-        field_set_list.append(acl_fieldset_entry)
-    field_set = ",".join(field_set_list) if field_set_list else None
-
-    # build filters
-    selection = f"{index_name}.document_id=='{chunk_request.document_id}'"
-    if chunk_request.min_chunk_ind is not None:
-        selection += f" and {index_name}.chunk_id>={chunk_request.min_chunk_ind}"
-    if chunk_request.max_chunk_ind is not None:
-        selection += f" and {index_name}.chunk_id<={chunk_request.max_chunk_ind}"
-    if not get_mega_chunks:
-        selection += f" and {index_name}.mega_chunk_reference_ids is empty"
-
-    # Setting up the selection criteria in the query parameters
-    params = {
-        # NOTE: Document Selector Language doesn't allow `contains`, so we can't check
-        # for the ACL in the selection. Instead, we have to check as a postfilter
-        "selection": selection,
-        "continuation": None,
-        "wantedDocumentCount": 1_000,
-        "fieldSet": field_set,
-    }
-
-    document_chunks: list[dict] = []
-    while True:
-        response = requests.get(url, params=params)
-        try:
-            response.raise_for_status()
-        except requests.HTTPError as e:
-            request_info = f"Headers: {response.request.headers}\nPayload: {params}"
-            response_info = f"Status Code: {response.status_code}\nResponse Content: {response.text}"
-            error_base = f"Error occurred getting chunk by Document ID {chunk_request.document_id}"
-            logger.error(
-                f"{error_base}:\n"
-                f"{request_info}\n"
-                f"{response_info}\n"
-                f"Exception: {e}"
-            )
-            raise requests.HTTPError(error_base) from e
-
-        # Check if the response contains any documents
-        response_data = response.json()
-        if "documents" in response_data:
-            for document in response_data["documents"]:
-                if user_access_control_list:
-                    document_acl = document["fields"].get(ACCESS_CONTROL_LIST)
-                    if not document_acl or not any(
-                        user_acl_entry in document_acl
-                        for user_acl_entry in user_access_control_list
-                    ):
-                        continue
-                document_chunks.append(document)
-
-        # Check for continuation token to handle pagination
-        if "continuation" in response_data and response_data["continuation"]:
-            params["continuation"] = response_data["continuation"]
-        else:
-            break  # Exit loop if no continuation token
-
-    return document_chunks
-
-
-def _get_all_vespa_chunk_ids_for_document_id(
-    document_id: str, index_name: str, user_access_control_list: list[str] | None = None
-) -> list[str]:
-    request = VespaChunkRequest(
-        document_id=document_id,
-        min_chunk_ind=None,
-        max_chunk_ind=None,
-    )
-    document_chunks = _get_vespa_chunks_by_document_id(
-        chunk_request=request,
-        index_name=index_name,
-        user_access_control_list=user_access_control_list,
-        field_names=[DOCUMENT_ID],
-    )
-    return [chunk["id"].split("::", 1)[-1] for chunk in document_chunks]
-
-
-def _parallelize_visit_api_retrieval(
-    index_name: str,
-    chunk_requests: list[VespaChunkRequest],
-    user_access_control_list: list[str] | None = None,
-) -> list[InferenceChunkUncleaned]:
-    functions_with_args: list[tuple[Callable, tuple]] = [
-        (
-            _get_vespa_chunks_by_document_id,
-            (chunk_request, index_name, user_access_control_list),
-        )
-        for chunk_request in chunk_requests
-    ]
-
-    parallel_results = run_functions_tuples_in_parallel(
-        functions_with_args, allow_failures=True
-    )
-
-    vespa_chunk_sets = [res for res in parallel_results if res]
-
-    flattened_vespa_chunks = [
-        chunk for chunk_set in vespa_chunk_sets for chunk in chunk_set
-    ]
-
-    inference_chunks = [
-        _vespa_hit_to_inference_chunk(chunk, null_score=True)
-        for chunk in flattened_vespa_chunks
-    ]
-
-    inference_chunks.sort(key=lambda chunk: chunk.chunk_id)
-
-    return inference_chunks
-
-
-def _build_vespa_id_based_retrieval_yql(request: VespaChunkRequest) -> str:
-    id_based_retrieval_yql_section = f"({DOCUMENT_ID} contains '{request.document_id}'"
-    if request.min_chunk_ind is not None:
-        id_based_retrieval_yql_section += f" and {request.min_chunk_ind} <= {CHUNK_ID}"
-    if request.max_chunk_ind is not None:
-        id_based_retrieval_yql_section += f" and {request.max_chunk_ind} >= {CHUNK_ID}"
-    id_based_retrieval_yql_section += ")"
-    return id_based_retrieval_yql_section
-
-
-def _batch_specified_chunk_retrieval(
-    index_name: str,
-    chunk_requests: list[VespaChunkRequest],
-    user_access_control_list: list[str] | None = None,
-) -> list[InferenceChunkUncleaned]:
-    filters = IndexFilters(access_control_list=None)
-    filters_str = _build_vespa_filters(filters=filters, include_hidden=True)
-
-    yql = (
-        VespaIndex.yql_base.format(index_name=index_name)
-        + filters_str
-        + _build_vespa_id_based_retrieval_yql(chunk_requests[0])
-    )
-    chunk_requests.pop(0)
-
-    for request in chunk_requests:
-        yql += " or " + _build_vespa_id_based_retrieval_yql(request)
-    params: dict[str, str | int | float] = {
-        "yql": yql,
-        "hits": _MAX_ID_SEARCH_QUERY_SIZE,
-    }
-
-    inference_chunks = _query_vespa(params)
-    inference_chunks.sort(key=lambda chunk: chunk.chunk_id)
-    return inference_chunks
-
-
-def _manage_batch_retrieval(
-    index_name: str,
-    chunk_requests: list[VespaChunkRequest],
-    user_access_control_list: list[str] | None = None,
-) -> list[InferenceChunkUncleaned]:
-    retrieved_chunks = []
-    capped_requests = []
-    uncapped_requests = []
-    chunk_count = 0
-    for request in chunk_requests:
-        # All requests without a chunk range are uncapped
-        # Uncapped requests are retrieved using the Visit API
-        if request.max_chunk_ind is None or request.min_chunk_ind is None:
-            uncapped_requests.append(request)
-            continue
-
-        # If the new addition to the chunk count is greater than the max query size,
-        # we need to perform a retrieval to avoid hitting the limit
-        new_addition = request.max_chunk_ind - request.min_chunk_ind + 1
-        if chunk_count + new_addition > _MAX_ID_SEARCH_QUERY_SIZE:
-            retrieved_chunks.extend(
-                _batch_specified_chunk_retrieval(
-                    index_name, capped_requests, user_access_control_list
-                )
-            )
-            capped_requests = []
-            chunk_count = 0
-        capped_requests.append(request)
-        chunk_count += new_addition
-
-    if capped_requests:
-        retrieved_chunks.extend(
-            _batch_specified_chunk_retrieval(
-                index_name, capped_requests, user_access_control_list
-            )
-        )
-
-    if uncapped_requests:
-        retrieved_chunks.extend(
-            _parallelize_visit_api_retrieval(
-                index_name, uncapped_requests, user_access_control_list
-            )
-        )
-    return retrieved_chunks
-
-
-@retry(tries=3, delay=1, backoff=2)
-def _delete_vespa_doc_chunks(
-    document_id: str, index_name: str, http_client: httpx.Client
-) -> None:
-    doc_chunk_ids = _get_all_vespa_chunk_ids_for_document_id(
-        document_id=document_id, index_name=index_name
-    )
-
-    for chunk_id in doc_chunk_ids:
-        try:
-            res = http_client.delete(
-                f"{DOCUMENT_ID_ENDPOINT.format(index_name=index_name)}/{chunk_id}"
-            )
-            res.raise_for_status()
-        except httpx.HTTPStatusError as e:
-            logger.error(f"Failed to delete chunk, details: {e.response.text}")
-            raise
-
-
-def _delete_vespa_docs(
-    document_ids: list[str],
-    index_name: str,
-    http_client: httpx.Client,
-    executor: concurrent.futures.ThreadPoolExecutor | None = None,
-) -> None:
-    external_executor = True
-
-    if not executor:
-        external_executor = False
-        executor = concurrent.futures.ThreadPoolExecutor(max_workers=_NUM_THREADS)
-
-    try:
-        doc_deletion_future = {
-            executor.submit(
-                _delete_vespa_doc_chunks, doc_id, index_name, http_client
-            ): doc_id
-            for doc_id in document_ids
-        }
-        for future in concurrent.futures.as_completed(doc_deletion_future):
-            # Will raise exception if the deletion raised an exception
-            future.result()
-
-    finally:
-        if not external_executor:
-            executor.shutdown(wait=True)
-
-
-def _get_existing_documents_from_chunks(
-    chunks: list[DocMetadataAwareIndexChunk],
-    index_name: str,
-    http_client: httpx.Client,
-    executor: concurrent.futures.ThreadPoolExecutor | None = None,
-) -> set[str]:
-    external_executor = True
-
-    if not executor:
-        external_executor = False
-        executor = concurrent.futures.ThreadPoolExecutor(max_workers=_NUM_THREADS)
-
-    document_ids: set[str] = set()
-    try:
-        chunk_existence_future = {
-            executor.submit(
-                _does_document_exist,
-                str(get_uuid_from_chunk(chunk)),
-                index_name,
-                http_client,
-            ): chunk
-            for chunk in chunks
-        }
-        for future in concurrent.futures.as_completed(chunk_existence_future):
-            chunk = chunk_existence_future[future]
-            chunk_already_existed = future.result()
-            if chunk_already_existed:
-                document_ids.add(chunk.source_document.id)
-
-    finally:
-        if not external_executor:
-            executor.shutdown(wait=True)
-
-    return document_ids
-
-
-@retry(tries=3, delay=1, backoff=2)
-def _index_vespa_chunk(
-    chunk: DocMetadataAwareIndexChunk, index_name: str, http_client: httpx.Client
-) -> None:
-    json_header = {
-        "Content-Type": "application/json",
-    }
-    document = chunk.source_document
-
-    # No minichunk documents in vespa, minichunk vectors are stored in the chunk itself
-    vespa_chunk_id = str(get_uuid_from_chunk(chunk))
-    embeddings = chunk.embeddings
-
-    embeddings_name_vector_map = {"full_chunk": embeddings.full_embedding}
-
-    if embeddings.mini_chunk_embeddings:
-        for ind, m_c_embed in enumerate(embeddings.mini_chunk_embeddings):
-            embeddings_name_vector_map[f"mini_chunk_{ind}"] = m_c_embed
-
-    title = document.get_title_for_document_index()
-
-    vespa_document_fields = {
-        DOCUMENT_ID: document.id,
-        CHUNK_ID: chunk.chunk_id,
-        BLURB: remove_invalid_unicode_chars(chunk.blurb),
-        TITLE: remove_invalid_unicode_chars(title) if title else None,
-        SKIP_TITLE_EMBEDDING: not title,
-        # For the BM25 index, the keyword suffix is used, the vector is already generated with the more
-        # natural language representation of the metadata section
-        CONTENT: remove_invalid_unicode_chars(
-            f"{chunk.title_prefix}{chunk.content}{chunk.metadata_suffix_keyword}"
-        ),
-        # This duplication of `content` is needed for keyword highlighting
-        # Note that it's not exactly the same as the actual content
-        # which contains the title prefix and metadata suffix
-        CONTENT_SUMMARY: remove_invalid_unicode_chars(chunk.content),
-        SOURCE_TYPE: str(document.source.value),
-        SOURCE_LINKS: json.dumps(chunk.source_links),
-        SEMANTIC_IDENTIFIER: remove_invalid_unicode_chars(document.semantic_identifier),
-        SECTION_CONTINUATION: chunk.section_continuation,
-        MEGA_CHUNK_REFERENCE_IDS: chunk.mega_chunk_reference_ids,
-        METADATA: json.dumps(document.metadata),
-        # Save as a list for efficient extraction as an Attribute
-        METADATA_LIST: chunk.source_document.get_metadata_str_attributes(),
-        METADATA_SUFFIX: chunk.metadata_suffix_keyword,
-        EMBEDDINGS: embeddings_name_vector_map,
-        TITLE_EMBEDDING: chunk.title_embedding,
-        BOOST: chunk.boost,
-        DOC_UPDATED_AT: _vespa_get_updated_at_attribute(document.doc_updated_at),
-        PRIMARY_OWNERS: get_experts_stores_representations(document.primary_owners),
-        SECONDARY_OWNERS: get_experts_stores_representations(document.secondary_owners),
-        # the only `set` vespa has is `weightedset`, so we have to give each
-        # element an arbitrary weight
-        ACCESS_CONTROL_LIST: {acl_entry: 1 for acl_entry in chunk.access.to_acl()},
-        DOCUMENT_SETS: {document_set: 1 for document_set in chunk.document_sets},
-    }
-
-    vespa_url = f"{DOCUMENT_ID_ENDPOINT.format(index_name=index_name)}/{vespa_chunk_id}"
-    logger.debug(f'Indexing to URL "{vespa_url}"')
-    res = http_client.post(
-        vespa_url, headers=json_header, json={"fields": vespa_document_fields}
-    )
-    try:
-        res.raise_for_status()
-    except Exception as e:
-        logger.exception(
-            f"Failed to index document: '{document.id}'. Got response: '{res.text}'"
-        )
-        raise e
-
-
-def _batch_index_vespa_chunks(
-    chunks: list[DocMetadataAwareIndexChunk],
-    index_name: str,
-    http_client: httpx.Client,
-    executor: concurrent.futures.ThreadPoolExecutor | None = None,
-) -> None:
-    external_executor = True
-
-    if not executor:
-        external_executor = False
-        executor = concurrent.futures.ThreadPoolExecutor(max_workers=_NUM_THREADS)
-
-    try:
-        chunk_index_future = {
-            executor.submit(_index_vespa_chunk, chunk, index_name, http_client): chunk
-            for chunk in chunks
-        }
-        for future in concurrent.futures.as_completed(chunk_index_future):
-            # Will raise exception if any indexing raised an exception
-            future.result()
-
-    finally:
-        if not external_executor:
-            executor.shutdown(wait=True)
-
-
-def _clear_and_index_vespa_chunks(
-    chunks: list[DocMetadataAwareIndexChunk],
-    index_name: str,
-) -> set[DocumentInsertionRecord]:
-    """Receive a list of chunks from a batch of documents and index the chunks into Vespa along
-    with updating the associated permissions. Assumes that a document will not be split into
-    multiple chunk batches calling this function multiple times, otherwise only the last set of
-    chunks will be kept"""
-    existing_docs: set[str] = set()
-
-    # NOTE: using `httpx` here since `requests` doesn't support HTTP2. This is beneficial for
-    # indexing / updates / deletes since we have to make a large volume of requests.
-    with (
-        concurrent.futures.ThreadPoolExecutor(max_workers=_NUM_THREADS) as executor,
-        httpx.Client(http2=True) as http_client,
-    ):
-        # Check for existing documents, existing documents need to have all of their chunks deleted
-        # prior to indexing as the document size (num chunks) may have shrunk
-        first_chunks = [chunk for chunk in chunks if chunk.chunk_id == 0]
-        for chunk_batch in batch_generator(first_chunks, _BATCH_SIZE):
-            existing_docs.update(
-                _get_existing_documents_from_chunks(
-                    chunks=chunk_batch,
-                    index_name=index_name,
-                    http_client=http_client,
-                    executor=executor,
-                )
-            )
-
-        for doc_id_batch in batch_generator(existing_docs, _BATCH_SIZE):
-            _delete_vespa_docs(
-                document_ids=doc_id_batch,
-                index_name=index_name,
-                http_client=http_client,
-                executor=executor,
-            )
-
-        for chunk_batch in batch_generator(chunks, _BATCH_SIZE):
-            _batch_index_vespa_chunks(
-                chunks=chunk_batch,
-                index_name=index_name,
-                http_client=http_client,
-                executor=executor,
-            )
-
-    all_doc_ids = {chunk.source_document.id for chunk in chunks}
-
-    return {
-        DocumentInsertionRecord(
-            document_id=doc_id,
-            already_existed=doc_id in existing_docs,
-        )
-        for doc_id in all_doc_ids
-    }
-
-
-def _build_vespa_filters(filters: IndexFilters, include_hidden: bool = False) -> str:
-    def _build_or_filters(key: str, vals: list[str] | None) -> str:
-        if vals is None:
-            return ""
-
-        valid_vals = [val for val in vals if val]
-        if not key or not valid_vals:
-            return ""
-
-        eq_elems = [f'{key} contains "{elem}"' for elem in valid_vals]
-        or_clause = " or ".join(eq_elems)
-        return f"({or_clause}) and "
-
-    def _build_time_filter(
-        cutoff: datetime | None,
-        # Slightly over 3 Months, approximately 1 fiscal quarter
-        untimed_doc_cutoff: timedelta = timedelta(days=92),
-    ) -> str:
-        if not cutoff:
-            return ""
-
-        # For Documents that don't have an updated at, filter them out for queries asking for
-        # very recent documents (3 months) default. Documents that don't have an updated at
-        # time are assigned 3 months for time decay value
-        include_untimed = datetime.now(timezone.utc) - untimed_doc_cutoff > cutoff
-        cutoff_secs = int(cutoff.timestamp())
-
-        if include_untimed:
-            # Documents without updated_at are assigned -1 as their date
-            return f"!({DOC_UPDATED_AT} < {cutoff_secs}) and "
-
-        return f"({DOC_UPDATED_AT} >= {cutoff_secs}) and "
-
-    filter_str = f"!({HIDDEN}=true) and " if not include_hidden else ""
-
-    # CAREFUL touching this one, currently there is no second ACL double-check post retrieval
-    if filters.access_control_list is not None:
-        filter_str += _build_or_filters(
-            ACCESS_CONTROL_LIST, filters.access_control_list
-        )
-
-    source_strs = (
-        [s.value for s in filters.source_type] if filters.source_type else None
-    )
-    filter_str += _build_or_filters(SOURCE_TYPE, source_strs)
-
-    tag_attributes = None
-    tags = filters.tags
-    if tags:
-        tag_attributes = [tag.tag_key + INDEX_SEPARATOR + tag.tag_value for tag in tags]
-    filter_str += _build_or_filters(METADATA_LIST, tag_attributes)
-
-    filter_str += _build_or_filters(DOCUMENT_SETS, filters.document_set)
-
-    filter_str += _build_time_filter(filters.time_cutoff)
-
-    return filter_str
-
-
-def _process_dynamic_summary(
-    dynamic_summary: str, max_summary_length: int = 400
-) -> list[str]:
-    if not dynamic_summary:
-        return []
-
-    current_length = 0
-    processed_summary: list[str] = []
-    for summary_section in dynamic_summary.split("<sep />"):
-        # if we're past the desired max length, break at the last word
-        if current_length + len(summary_section) >= max_summary_length:
-            summary_section = summary_section[: max_summary_length - current_length]
-            summary_section = summary_section.lstrip()  # remove any leading whitespace
-
-            # handle the case where the truncated section is either just a
-            # single (partial) word or if it's empty
-            first_space = summary_section.find(" ")
-            if first_space == -1:
-                # add ``...`` to previous section
-                if processed_summary:
-                    processed_summary[-1] += "..."
-                break
-
-            # handle the valid truncated section case
-            summary_section = summary_section.rsplit(" ", 1)[0]
-            if summary_section[-1] in string.punctuation:
-                summary_section = summary_section[:-1]
-            summary_section += "..."
-            processed_summary.append(summary_section)
-            break
-
-        processed_summary.append(summary_section)
-        current_length += len(summary_section)
-
-    return processed_summary
-
-
-def _vespa_hit_to_inference_chunk(
-    hit: dict[str, Any], null_score: bool = False
-) -> InferenceChunkUncleaned:
-    fields = cast(dict[str, Any], hit["fields"])
-
-    # parse fields that are stored as strings, but are really json / datetime
-    metadata = json.loads(fields[METADATA]) if METADATA in fields else {}
-    updated_at = (
-        datetime.fromtimestamp(fields[DOC_UPDATED_AT], tz=timezone.utc)
-        if DOC_UPDATED_AT in fields
-        else None
-    )
-
-    # The highlights might include the title but this is the best way we have so far to show the highlighting
-    match_highlights = _process_dynamic_summary(
-        # fallback to regular `content` if the `content_summary` field
-        # isn't present
-        dynamic_summary=hit["fields"].get(CONTENT_SUMMARY, hit["fields"][CONTENT]),
-    )
-    semantic_identifier = fields.get(SEMANTIC_IDENTIFIER, "")
-    if not semantic_identifier:
-        logger.error(
-            f"Chunk with blurb: {fields.get(BLURB, 'Unknown')[:50]}... has no Semantic Identifier"
-        )
-
-    source_links = fields.get(SOURCE_LINKS, {})
-    source_links_dict_unprocessed = (
-        json.loads(source_links) if isinstance(source_links, str) else source_links
-    )
-    source_links_dict = {
-        int(k): v
-        for k, v in cast(dict[str, str], source_links_dict_unprocessed).items()
-    }
-
-    return InferenceChunkUncleaned(
-        chunk_id=fields[CHUNK_ID],
-        blurb=fields.get(BLURB, ""),  # Unused
-        content=fields[CONTENT],  # Includes extra title prefix and metadata suffix
-        source_links=source_links_dict,
-        section_continuation=fields[SECTION_CONTINUATION],
-        document_id=fields[DOCUMENT_ID],
-        source_type=fields[SOURCE_TYPE],
-        title=fields.get(TITLE),
-        semantic_identifier=fields[SEMANTIC_IDENTIFIER],
-        boost=fields.get(BOOST, 1),
-        recency_bias=fields.get("matchfeatures", {}).get(RECENCY_BIAS, 1.0),
-        score=None if null_score else hit.get("relevance", 0),
-        hidden=fields.get(HIDDEN, False),
-        primary_owners=fields.get(PRIMARY_OWNERS),
-        secondary_owners=fields.get(SECONDARY_OWNERS),
-        mega_chunk_reference_ids=fields.get(MEGA_CHUNK_REFERENCE_IDS, []),
-        metadata=metadata,
-        metadata_suffix=fields.get(METADATA_SUFFIX),
-        match_highlights=match_highlights,
-        updated_at=updated_at,
-    )
-
-
-@retry(tries=3, delay=1, backoff=2)
-def _query_vespa(
-    query_params: Mapping[str, str | int | float]
-) -> list[InferenceChunkUncleaned]:
-    if "query" in query_params and not cast(str, query_params["query"]).strip():
-        raise ValueError("No/empty query received")
-
-    params = dict(
-        **query_params,
-        **{
-            "presentation.timing": True,
-        }
-        if LOG_VESPA_TIMING_INFORMATION
-        else {},
-    )
-
-    response = requests.post(
-        SEARCH_ENDPOINT,
-        json=params,
-    )
-    try:
-        response.raise_for_status()
-    except requests.HTTPError as e:
-        request_info = f"Headers: {response.request.headers}\nPayload: {params}"
-        response_info = (
-            f"Status Code: {response.status_code}\n"
-            f"Response Content: {response.text}"
-        )
-        error_base = "Failed to query Vespa"
-        logger.error(
-            f"{error_base}:\n"
-            f"{request_info}\n"
-            f"{response_info}\n"
-            f"Exception: {e}"
-        )
-        raise requests.HTTPError(error_base) from e
-
-    response_json: dict[str, Any] = response.json()
-    if LOG_VESPA_TIMING_INFORMATION:
-        logger.info("Vespa timing info: %s", response_json.get("timing"))
-    hits = response_json["root"].get("children", [])
-
-    for hit in hits:
-        if hit["fields"].get(CONTENT) is None:
-            identifier = hit["fields"].get("documentid") or hit["id"]
-            logger.error(
-                f"Vespa Index with Vespa ID {identifier} has no contents. "
-                f"This is invalid because the vector is not meaningful and keywordsearch cannot "
-                f"fetch this document"
-            )
-
-    filtered_hits = [hit for hit in hits if hit["fields"].get(CONTENT) is not None]
-
-    inference_chunks = [_vespa_hit_to_inference_chunk(hit) for hit in filtered_hits]
-    # Good Debugging Spot
-    return inference_chunks
 
 
 def in_memory_zip_from_file_bytes(file_contents: dict[str, bytes]) -> BinaryIO:
@@ -836,46 +91,7 @@ def _create_document_xml_lines(doc_names: list[str | None]) -> str:
     return "\n".join(doc_lines)
 
 
-def _clean_chunk_id_copy(
-    chunk: DocMetadataAwareIndexChunk,
-) -> DocMetadataAwareIndexChunk:
-    clean_chunk = chunk.copy(
-        update={
-            "source_document": chunk.source_document.copy(
-                update={
-                    "id": replace_invalid_doc_id_characters(chunk.source_document.id)
-                }
-            )
-        }
-    )
-    return clean_chunk
-
-
 class VespaIndex(DocumentIndex):
-    yql_base = (
-        f"select "
-        f"documentid, "
-        f"{DOCUMENT_ID}, "
-        f"{CHUNK_ID}, "
-        f"{BLURB}, "
-        f"{CONTENT}, "
-        f"{SOURCE_TYPE}, "
-        f"{SOURCE_LINKS}, "
-        f"{SEMANTIC_IDENTIFIER}, "
-        f"{TITLE}, "
-        f"{SECTION_CONTINUATION}, "
-        f"{BOOST}, "
-        f"{HIDDEN}, "
-        f"{DOC_UPDATED_AT}, "
-        f"{PRIMARY_OWNERS}, "
-        f"{SECONDARY_OWNERS}, "
-        f"{MEGA_CHUNK_REFERENCE_IDS}, "
-        f"{METADATA}, "
-        f"{METADATA_SUFFIX}, "
-        f"{CONTENT_SUMMARY} "
-        f"from {{index_name}} where "
-    )
-
     def __init__(self, index_name: str, secondary_index_name: str | None) -> None:
         self.index_name = index_name
         self.secondary_index_name = secondary_index_name
@@ -946,16 +162,64 @@ class VespaIndex(DocumentIndex):
         self,
         chunks: list[DocMetadataAwareIndexChunk],
     ) -> set[DocumentInsertionRecord]:
+        """Receive a list of chunks from a batch of documents and index the chunks into Vespa along
+        with updating the associated permissions. Assumes that a document will not be split into
+        multiple chunk batches calling this function multiple times, otherwise only the last set of
+        chunks will be kept"""
         # IMPORTANT: This must be done one index at a time, do not use secondary index here
-        cleaned_chunks = [_clean_chunk_id_copy(chunk) for chunk in chunks]
-        return _clear_and_index_vespa_chunks(
-            chunks=cleaned_chunks, index_name=self.index_name
-        )
+        cleaned_chunks = [clean_chunk_id_copy(chunk) for chunk in chunks]
+
+        existing_docs: set[str] = set()
+
+        # NOTE: using `httpx` here since `requests` doesn't support HTTP2. This is beneficial for
+        # indexing / updates / deletes since we have to make a large volume of requests.
+        with (
+            concurrent.futures.ThreadPoolExecutor(max_workers=NUM_THREADS) as executor,
+            httpx.Client(http2=True) as http_client,
+        ):
+            # Check for existing documents, existing documents need to have all of their chunks deleted
+            # prior to indexing as the document size (num chunks) may have shrunk
+            first_chunks = [chunk for chunk in cleaned_chunks if chunk.chunk_id == 0]
+            for chunk_batch in batch_generator(first_chunks, BATCH_SIZE):
+                existing_docs.update(
+                    get_existing_documents_from_chunks(
+                        chunks=chunk_batch,
+                        index_name=self.index_name,
+                        http_client=http_client,
+                        executor=executor,
+                    )
+                )
+
+            for doc_id_batch in batch_generator(existing_docs, BATCH_SIZE):
+                delete_vespa_docs(
+                    document_ids=doc_id_batch,
+                    index_name=self.index_name,
+                    http_client=http_client,
+                    executor=executor,
+                )
+
+            for chunk_batch in batch_generator(cleaned_chunks, BATCH_SIZE):
+                batch_index_vespa_chunks(
+                    chunks=chunk_batch,
+                    index_name=self.index_name,
+                    http_client=http_client,
+                    executor=executor,
+                )
+
+        all_doc_ids = {chunk.source_document.id for chunk in cleaned_chunks}
+
+        return {
+            DocumentInsertionRecord(
+                document_id=doc_id,
+                already_existed=doc_id in existing_docs,
+            )
+            for doc_id in all_doc_ids
+        }
 
     @staticmethod
     def _apply_updates_batched(
         updates: list[_VespaUpdateRequest],
-        batch_size: int = _BATCH_SIZE,
+        batch_size: int = BATCH_SIZE,
     ) -> None:
         """Runs a batch of updates in parallel via the ThreadPoolExecutor."""
 
@@ -974,7 +238,7 @@ class VespaIndex(DocumentIndex):
         # NOTE: using `httpx` here since `requests` doesn't support HTTP2. This is beneficient for
         # indexing / updates / deletes since we have to make a large volume of requests.
         with (
-            concurrent.futures.ThreadPoolExecutor(max_workers=_NUM_THREADS) as executor,
+            concurrent.futures.ThreadPoolExecutor(max_workers=NUM_THREADS) as executor,
             httpx.Client(http2=True) as http_client,
         ):
             for update_batch in batch_generator(updates, batch_size):
@@ -1016,14 +280,14 @@ class VespaIndex(DocumentIndex):
             index_names.append(self.secondary_index_name)
 
         chunk_id_start_time = time.monotonic()
-        with concurrent.futures.ThreadPoolExecutor(
-            max_workers=_NUM_THREADS
-        ) as executor:
+        with concurrent.futures.ThreadPoolExecutor(max_workers=NUM_THREADS) as executor:
             future_to_doc_chunk_ids = {
                 executor.submit(
-                    _get_all_vespa_chunk_ids_for_document_id,
+                    get_all_vespa_ids_for_document_id,
                     document_id=document_id,
                     index_name=index_name,
+                    filters=None,
+                    get_mega_chunks=False,
                 ): (document_id, index_name)
                 for index_name in index_names
                 for update_request in update_requests
@@ -1097,26 +361,29 @@ class VespaIndex(DocumentIndex):
                 index_names.append(self.secondary_index_name)
 
             for index_name in index_names:
-                _delete_vespa_docs(
+                delete_vespa_docs(
                     document_ids=doc_ids, index_name=index_name, http_client=http_client
                 )
 
     def id_based_retrieval(
         self,
         chunk_requests: list[VespaChunkRequest],
-        user_access_control_list: list[str] | None = None,
+        filters: IndexFilters,
         batch_retrieval: bool = False,
+        get_mega_chunks: bool = False,
     ) -> list[InferenceChunkUncleaned]:
         if batch_retrieval:
-            return _manage_batch_retrieval(
-                self.index_name,
-                chunk_requests,
-                user_access_control_list,
+            return manage_batch_retrieval(
+                index_name=self.index_name,
+                chunk_requests=chunk_requests,
+                filters=filters,
+                get_mega_chunks=get_mega_chunks,
             )
-        return _parallelize_visit_api_retrieval(
-            self.index_name,
-            chunk_requests,
-            user_access_control_list,
+        return parallel_visit_api_retrieval(
+            index_name=self.index_name,
+            chunk_requests=chunk_requests,
+            filters=filters,
+            get_mega_chunks=get_mega_chunks,
         )
 
     def keyword_retrieval(
@@ -1129,9 +396,9 @@ class VespaIndex(DocumentIndex):
         edit_keyword_query: bool = EDIT_KEYWORD_QUERY,
     ) -> list[InferenceChunkUncleaned]:
         # IMPORTANT: THIS FUNCTION IS NOT UP TO DATE, DOES NOT WORK CORRECTLY
-        vespa_where_clauses = _build_vespa_filters(filters)
+        vespa_where_clauses = build_vespa_filters(filters)
         yql = (
-            VespaIndex.yql_base.format(index_name=self.index_name)
+            YQL_BASE.format(index_name=self.index_name)
             + vespa_where_clauses
             # `({defaultIndex: "content_summary"}userInput(@query))` section is
             # needed for highlighting while the N-gram highlighting is broken /
@@ -1149,10 +416,10 @@ class VespaIndex(DocumentIndex):
             "hits": num_to_retrieve,
             "offset": offset,
             "ranking.profile": "keyword_search",
-            "timeout": _VESPA_TIMEOUT,
+            "timeout": VESPA_TIMEOUT,
         }
 
-        return _query_vespa(params)
+        return query_vespa(params)
 
     def semantic_retrieval(
         self,
@@ -1166,9 +433,9 @@ class VespaIndex(DocumentIndex):
         edit_keyword_query: bool = EDIT_KEYWORD_QUERY,
     ) -> list[InferenceChunkUncleaned]:
         # IMPORTANT: THIS FUNCTION IS NOT UP TO DATE, DOES NOT WORK CORRECTLY
-        vespa_where_clauses = _build_vespa_filters(filters)
+        vespa_where_clauses = build_vespa_filters(filters)
         yql = (
-            VespaIndex.yql_base.format(index_name=self.index_name)
+            YQL_BASE.format(index_name=self.index_name)
             + vespa_where_clauses
             + f"(({{targetHits: {10 * num_to_retrieve}}}nearestNeighbor(embeddings, query_embedding)) "
             # `({defaultIndex: "content_summary"}userInput(@query))` section is
@@ -1191,10 +458,10 @@ class VespaIndex(DocumentIndex):
             "hits": num_to_retrieve,
             "offset": offset,
             "ranking.profile": f"hybrid_search{len(query_embedding)}",
-            "timeout": _VESPA_TIMEOUT,
+            "timeout": VESPA_TIMEOUT,
         }
 
-        return _query_vespa(params)
+        return query_vespa(params)
 
     def hybrid_retrieval(
         self,
@@ -1209,11 +476,11 @@ class VespaIndex(DocumentIndex):
         distance_cutoff: float | None = SEARCH_DISTANCE_CUTOFF,
         edit_keyword_query: bool = EDIT_KEYWORD_QUERY,
     ) -> list[InferenceChunkUncleaned]:
-        vespa_where_clauses = _build_vespa_filters(filters)
+        vespa_where_clauses = build_vespa_filters(filters)
         # Needs to be at least as much as the value set in Vespa schema config
         target_hits = max(10 * num_to_retrieve, 1000)
         yql = (
-            VespaIndex.yql_base.format(index_name=self.index_name)
+            YQL_BASE.format(index_name=self.index_name)
             + vespa_where_clauses
             + f"(({{targetHits: {target_hits}}}nearestNeighbor(embeddings, query_embedding)) "
             + f"or ({{targetHits: {target_hits}}}nearestNeighbor(title_embedding, query_embedding)) "
@@ -1241,10 +508,10 @@ class VespaIndex(DocumentIndex):
             "hits": num_to_retrieve,
             "offset": offset,
             "ranking.profile": f"hybrid_search{len(query_embedding)}",
-            "timeout": _VESPA_TIMEOUT,
+            "timeout": VESPA_TIMEOUT,
         }
 
-        return _query_vespa(params)
+        return query_vespa(params)
 
     def admin_retrieval(
         self,
@@ -1253,9 +520,9 @@ class VespaIndex(DocumentIndex):
         num_to_retrieve: int = NUM_RETURNED_HITS,
         offset: int = 0,
     ) -> list[InferenceChunkUncleaned]:
-        vespa_where_clauses = _build_vespa_filters(filters, include_hidden=True)
+        vespa_where_clauses = build_vespa_filters(filters, include_hidden=True)
         yql = (
-            VespaIndex.yql_base.format(index_name=self.index_name)
+            YQL_BASE.format(index_name=self.index_name)
             + vespa_where_clauses
             + '({grammar: "weakAnd"}userInput(@query) '
             # `({defaultIndex: "content_summary"}userInput(@query))` section is
@@ -1270,7 +537,7 @@ class VespaIndex(DocumentIndex):
             "hits": num_to_retrieve,
             "offset": 0,
             "ranking.profile": "admin_search",
-            "timeout": _VESPA_TIMEOUT,
+            "timeout": VESPA_TIMEOUT,
         }
 
-        return _query_vespa(params)
+        return query_vespa(params)
