+++ backend/danswer/server/documents/document.py
@@ -9,8 +9,10 @@ from danswer.db.embedding_model import get_current_db_embedding_model
 from danswer.db.engine import get_session
 from danswer.db.models import User
 from danswer.document_index.factory import get_default_document_index
+from danswer.document_index.interfaces import VespaChunkRequest
 from danswer.natural_language_processing.utils import get_tokenizer
 from danswer.prompts.prompt_utils import build_doc_context_str
+from danswer.search.models import IndexFilters
 from danswer.search.preprocessing.access_filters import build_access_filters_for_user
 from danswer.server.documents.models import ChunkInfo
 from danswer.server.documents.models import DocumentInfo
@@ -35,10 +37,8 @@ def get_document_info(
 
     user_acl_filters = build_access_filters_for_user(user, db_session)
     inference_chunks = document_index.id_based_retrieval(
-        document_id=document_id,
-        min_chunk_ind=None,
-        max_chunk_ind=None,
-        user_access_control_list=user_acl_filters,
+        chunk_requests=[VespaChunkRequest(document_id=document_id)],
+        filters=IndexFilters(access_control_list=user_acl_filters),
     )
 
     if not inference_chunks:
@@ -83,11 +83,15 @@ def get_chunk_info(
     )
 
     user_acl_filters = build_access_filters_for_user(user, db_session)
-    inference_chunks = document_index.id_based_retrieval(
+    chunk_request = VespaChunkRequest(
         document_id=document_id,
         min_chunk_ind=chunk_id,
         max_chunk_ind=chunk_id,
-        user_access_control_list=user_acl_filters,
+    )
+    inference_chunks = document_index.id_based_retrieval(
+        chunk_requests=[chunk_request],
+        filters=IndexFilters(access_control_list=user_acl_filters),
+        batch_retrieval=True,
     )
 
     if not inference_chunks:
